Auto stash before merge of "feature/CLX-382-claims-graphql" and "origin/feature/CLX-382-claims-graphql"
Added seperatedModel for phoneStep
